Moved abstract classes and other provider/model implementations to their domain specific directories.
