+++ src/Providers/ApiBasedImplementation/AbstractApiBasedModel.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace WordPress\AiClient\Providers\Models;
+namespace WordPress\AiClient\Providers\ApiBasedImplementation;
 
 use WordPress\AiClient\Providers\DTO\ProviderMetadata;
 use WordPress\AiClient\Providers\Http\Contracts\WithHttpTransporterInterface;
