+++ src/Providers/ApiBasedImplementation/AbstractApiBasedModelMetadataDirectory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace WordPress\AiClient\Providers;
+namespace WordPress\AiClient\Providers\ApiBasedImplementation;
 
 use InvalidArgumentException;
 use WordPress\AiClient\Providers\Contracts\ModelMetadataDirectoryInterface;
